update count order in folder
